requirements.txt
Flask>=2.0
pandas
Pillow
qrcode
